# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MTech local AI agent workspace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MtechHealth {
  status: string;
  service: string;
  version: string;
}

export interface LlmConfig {
  provider: string;
  endpoint: string;
  model: string;
  apiKeyConfigured: boolean;
  timeout: number;
  maxTokens: number;
  temperature: number;
}

export interface LlmConfigInput {
  endpoint?: string;
  model?: string;
  timeout?: number;
  maxTokens?: number;
  temperature?: number;
}

export type LlmStatus = LlmConfig & ({
  connected: boolean;
  /** @nullable */
  latency: number | null;
  /** @nullable */
  lastCheck: string | null;
  /** @nullable */
  error: string | null;
});

export interface Model {
  id: string;
  ownedBy: string;
}

export interface ModelList {
  models: Model[];
}

export interface LlmTestResult {
  connected: boolean;
  /** @nullable */
  latency: number | null;
  message: string;
  model: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatInput {
  projectId: string;
  /** @minLength 1 */
  message: string;
  history?: ChatMessage[];
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  agent: 'agent',
  system: 'system',
  success: 'success',
  warning: 'warning',
} as const;

export interface ActivityEvent {
  id: string;
  agent: string;
  action: string;
  detail: string;
  timestamp: string;
  type: ActivityEventType;
}

export interface ChatResponse {
  message: string;
  taskId: string;
  activity: ActivityEvent[];
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  paused: 'paused',
  archived: 'archived',
} as const;

export type ProjectFileKind = typeof ProjectFileKind[keyof typeof ProjectFileKind];


export const ProjectFileKind = {
  file: 'file',
  folder: 'folder',
} as const;

export interface ProjectFile {
  path: string;
  kind: ProjectFileKind;
  language: string;
  size: number;
  modified: boolean;
}

export interface Project {
  id: string;
  name: string;
  description: string;
  status: ProjectStatus;
  branch: string;
  updatedAt: string;
  files: ProjectFile[];
  progress: number;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  running: 'running',
  complete: 'complete',
  queued: 'queued',
  blocked: 'blocked',
} as const;

export interface Agent {
  id: string;
  name: string;
  role: string;
  status: AgentStatus;
  currentTask: string;
  progress: number;
  lastAction: string;
  changedFiles: number;
}

export type AgentTaskStatus = typeof AgentTaskStatus[keyof typeof AgentTaskStatus];


export const AgentTaskStatus = {
  queued: 'queued',
  planning: 'planning',
  running: 'running',
  waiting: 'waiting',
  completed: 'completed',
  failed: 'failed',
  retrying: 'retrying',
  cancelled: 'cancelled',
} as const;

export type AgentTaskPriority = typeof AgentTaskPriority[keyof typeof AgentTaskPriority];


export const AgentTaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface AgentTask {
  id: string;
  projectId: string;
  agentId: string;
  title: string;
  description: string;
  status: AgentTaskStatus;
  priority: AgentTaskPriority;
  dependencies: string[];
  createdAt: string;
  attempts: number;
}

export type PreviewStatusStatus = typeof PreviewStatusStatus[keyof typeof PreviewStatusStatus];


export const PreviewStatusStatus = {
  running: 'running',
  stopped: 'stopped',
  starting: 'starting',
} as const;

export interface PreviewStatus {
  status: PreviewStatusStatus;
  /** @nullable */
  url: string | null;
  /** @nullable */
  port: number | null;
}

export type DownloadProjectFileParams = {
/**
 * Workspace-relative file path.
 */
path: string;
};