# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MTech local AI agent workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventType } from './activityEventType';

export interface ActivityEvent {
  id: string;
  agent: string;
  action: string;
  detail: string;
  timestamp: Date;
  type: ActivityEventType;
}