# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MTech local AI agent workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatInput {
  projectId: string;
  /** @minLength 1 */
  message: string;
  history?: ChatMessage[];
}