# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MTech local AI agent workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEvent } from './activityEvent';

export interface ChatResponse {
  message: string;
  taskId: string;
  activity: ActivityEvent[];
}