# lib/api-zod/src/generated/types/llmConfig.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MTech local AI agent workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface LlmConfig {
  provider: string;
  endpoint: string;
  model: string;
  apiKeyConfigured: boolean;
  timeout: number;
  maxTokens: number;
  temperature: number;
}