# lib/api-zod/src/generated/types/model.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MTech local AI agent workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface Model {
  id: string;
  ownedBy: string;
}